Add result controller to admin module
