+++ app/views/application/_tools_menu.html.slim
@@ -6,5 +6,5 @@
     li
       = link_to("Categorise listing", admin_categories_url )
     li
-      = link_to("Add result", new_result_path)
+      = link_to("Add result", new_admin_result_path)
